requirements.txt
django>=4.2
django-environ>=0.11
twilio>=9.0
pandas>=2.0
openpyxl>=3.1
Pillow>=10.0
whitenoise>=6.6
gunicorn>=22.0
psycopg[binary]>=3.2
